+++ rmk/CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `report_hz` option for Pmw3610Device
+
 ### Changed
 
 - **BREAKING**: `PollingController::INTERVAL` constant is now `PollingController::interval()` method, allowing dynamic interval configuration at runtime
+- Optimize the timing for motion read and sending reports on the PMW3610
 
 ## [0.8.2] - 2025-12-18